% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{as.data.frame.plcp_multi_sim_summary}
\alias{as.data.frame.plcp_multi_sim_summary}
\title{Convert a multi-sim summary object to a tidy data.frame}
\usage{
\method{as.data.frame}{plcp_multi_sim_summary}(x, ...)
}
\arguments{
\item{x}{Object with class \code{plcp_multi_sim_summary}.}

\item{...}{Not used}
}
\value{
a \code{data.frame} with one row for each simulation.
Columns include the simulation study parameters and the results.
}
\description{
Convert a multi-sim summary object to a tidy data.frame
}
