% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{create_lmer_formula}
\alias{create_lmer_formula}
\title{Create an lmer formula based on a \code{\link{study_parameters}}-object}
\usage{
create_lmer_formula(object, ...)
}
\arguments{
\item{object}{A \code{\link{study_parameters}}-object containing one study design}

\item{...}{Unused, optional arguments.}
}
\value{
A \code{character} vector with lmer formula syntax.
}
\description{
Create an lmer formula based on a \code{\link{study_parameters}}-object
}
\details{
The lme4 formula will correspond to the model implied by the specified parameters in
the \code{\link{study_parameters}}-object. Thus, if e.g. \code{cor_subject} is \code{NA} or \code{NULL} the
corresponding term is removed from the lmer formula. Parameters that are 0 are retained.


Currently only objects with one study design are supported, i.e. objects with class \code{plcp},
and not \code{plcp_multi}; \code{data.frame} with multiple designs are currently not supported.
}
