% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdle_simulate_data.R
\name{create_R_cov}
\alias{create_R_cov}
\title{add effect sizes to paras
add function to plot by tx group}
\usage{
create_R_cov(pars)
}
\description{
add effect sizes to paras
add function to plot by tx group
}
