% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hurdle.R
\name{get_FE_thetas.plcp_hurdle}
\alias{get_FE_thetas.plcp_hurdle}
\title{#' @export
get_n3.plcp_hurdle <- function(paras) {
    NA
}}
\usage{
\method{get_FE_thetas}{plcp_hurdle}(paras, marginalize = FALSE, ...)
}
\description{
#' @export
get_n3.plcp_hurdle <- function(paras) {
    NA
}
}
