% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_ICC_pre_subjects}
\alias{get_ICC_pre_subjects}
\title{Calculates the amount of baseline variance at the subject level}
\usage{
get_ICC_pre_subjects(object, ...)
}
\arguments{
\item{object}{An object created by \code{\link{study_parameters}}}

\item{...}{Optional named arguments.}
}
\value{
Returns the proportion of baseline variance at the subject level,
as a numeric vector.
}
\description{
Calculates the amount of baseline variance at the subject level
}
\examples{
paras <- study_parameters(n1 = 11,
                          n2 = 10,
                          n3 = 3,
                          T_end = 10,
                          sigma_subject_intercept = 1.2,
                          sigma_subject_slope = 0.2,
                          sigma_cluster_intercept = 0.5,
                          sigma_cluster_slope = 0.2,
                          sigma_error = 1.2,
                          cohend = -0.8)

get_ICC_pre_subjects(paras)
}
