% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hurdle.R
\name{get_RE_thetas.plcp_hurdle}
\alias{get_RE_thetas.plcp_hurdle}
\title{#' @export
get_tot_n.plcp_hurdle <- function(paras) {
    paras$n2 * 2
}}
\usage{
\method{get_RE_thetas}{plcp_hurdle}(paras)
}
\description{
#' @export
get_tot_n.plcp_hurdle <- function(paras) {
    paras$n2 * 2
}
}
