% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_slope_diff}
\alias{get_slope_diff}
\title{add standardizers
* pooled pre or posttest
* control group SD
* random slope SD}
\usage{
get_slope_diff(paras)
}
\description{
add standardizers
* pooled pre or posttest
* control group SD
* random slope SD
}
