% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdle_marginalize.R
\name{marginalize}
\alias{marginalize}
\title{Marginalize over random effects and zero and positive part}
\usage{
marginalize(object, R, ...)
}
\arguments{
\item{object}{A \code{plcp_hurdle} object created
using \code{study_parameters}}

\item{R}{\code{integer} number of Monte Carlo replications}

\item{...}{}
}
\description{
Marginalize over random effects and zero and positive part
}
