% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPC.r
\name{plot.plcp_ICC2}
\alias{plot.plcp_ICC2}
\title{Plot method for \code{get_correlation_matrix}-objects}
\usage{
\method{plot}{plcp_ICC2}(x, ...)
}
\arguments{
\item{x}{An object created with \code{\link{get_correlation_matrix}}}

\item{...}{Optional arguments, currently ignored.}
}
\description{
Plot method for \code{get_correlation_matrix}-objects
}
