% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot.plcp_power_table}
\alias{plot.plcp_power_table}
\title{Plot method for \code{get_power_table}-objects}
\usage{
\method{plot}{plcp_power_table}(x, ...)
}
\arguments{
\item{x}{An object of class \code{plcp_power_table}.}

\item{...}{Optional arguments.}
}
\description{
Plot method for \code{get_power_table}-objects
}
