% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPC.r
\name{plot.plcp_sds}
\alias{plot.plcp_sds}
\title{Plot method for \code{get_sds}-objects}
\usage{
\method{plot}{plcp_sds}(x, ...)
}
\arguments{
\item{x}{An object of class \code{plcp_sds}.}

\item{...}{Optional arguments.}
}
\description{
Plot method for \code{get_sds}-objects
}
