% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPC.r
\name{plot.plcp_VPC}
\alias{plot.plcp_VPC}
\title{Plot method for \code{get_VPC}-objects}
\usage{
\method{plot}{plcp_VPC}(x, ...)
}
\arguments{
\item{x}{An object created with \code{\link{get_VPC}}}

\item{...}{Optional arguments, currently ignored.}
}
\description{
Plot method for \code{get_VPC}-objects
}
