% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlmm-package.R
\docType{package}
\name{powerlmm}
\alias{powerlmm}
\alias{powerlmm-package}
\title{powerlmm}
\description{
The \pkg{powerlmm} provides a fast and flexible way to calculate power for
two- and three-level multilevel models with missing data. The focus is on
longitudinal designs, i.e. where level 1 is measurements, level 2 subjects nested within a higher
level 3 unit, e.g. therapists.
}
\details{
All study designs are specified using the function \code{\link{study_parameters}},
which lets you define your model using familiar notation, either by specifying
the model parameters directly, or by using relative standardize inputs (e.g. % variance at
each level). Several functions are provide to help you visualize and understand
implied model, type \code{methods(class="plcp")} to see available methods.
}
\section{Supported models}{

The purpose of \pkg{powerlmm} is to help design longitudinal treatment studies,
with or without higher-level clustering (e.g. by therapists, groups, or physician),
and missing data. The main features of the package are:
\itemize{
\item (Longitudinal) Two- and three-level (nested) linear mixed models,
and partially nested designs
\item Random slopes at the subject- and cluster-level
\item Account for missing data/dropout
\item Unbalanced designs (both unequal cluster sizes, and treatment groups)
\item Calculate the design effect, and estimated type I error when the
third-level is ignored.
\item Fast analytical power calculations for all supported designs
\item Explore bias, type 1 error and model misspecification using
convenient simulation methods.
}
}

\section{Non-longitudinal models}{

The focus of this package is on longitudinal models. However, it would also
be possible to calculate power for studies with a similar structure, e.g. a
 three-level model with pupils in classes nested within schools.
}

\section{Tutorials}{

Type \code{vignette("two-level", package = "powerlmm")}, or
\code{vignette("three-level", package = "powerlmm")} to see a tutorial on
using \pkg{powerlmm} to calculate power. See all available vignettes by typing
\code{vignette(package = "powerlmm")}.
}

\seealso{
\code{\link{study_parameters}}, \code{\link{get_power}}
}
\author{
Kristoffer Magnusson

Maintainer: Kristoffer Magnusson <hello@kristoffer.email>
}
