% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{print.plcp_2lvl}
\alias{print.plcp_2lvl}
\title{Print method for two-level \code{study_parameters}-objects}
\usage{
\method{print}{plcp_2lvl}(x, ...)
}
\arguments{
\item{x}{An object of class \code{plcp_2lvl}.}

\item{...}{Optional arguments.}
}
\description{
Print method for two-level \code{study_parameters}-objects
}
