% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{print.plcp_3lvl}
\alias{print.plcp_3lvl}
\title{Print method for three-level \code{study_parameters}-objects}
\usage{
\method{print}{plcp_3lvl}(x, ...)
}
\arguments{
\item{x}{An object of class \code{plcp_3lvl}.}

\item{...}{Optional arguments.}
}
\description{
Print method for three-level \code{study_parameters}-objects
}
