% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_print.R
\name{print.plcp_hurdle}
\alias{print.plcp_hurdle}
\title{Print method for two-part/hurdle \code{study_parameters}-objects}
\usage{
\method{print}{plcp_2lvl}(x, ...)
}
\arguments{
\item{x}{An object of class \code{plcp_hurdle}.}

\item{...}{Optional arguments.}
}
\description{
Print method for two-part/hurdle \code{study_parameters}-objects
}
