% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPC.r
\name{print.plcp_ICC2}
\alias{print.plcp_ICC2}
\title{Print method for \code{get_correlation_matrix}-objects}
\usage{
\method{print}{plcp_ICC2}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object created by \code{\link{get_correlation_matrix}}}

\item{...}{Optional arguments}
}
\description{
Print method for \code{get_correlation_matrix}-objects
}
