% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print.plcp_mc_se}
\alias{print.plcp_mc_se}
\title{Print method for \code{get_monte_carlo_se}-objects}
\usage{
\method{print}{plcp_mc_se}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object created with \code{\link{get_monte_carlo_se}}.}

\item{digits}{The number of digits to print.}

\item{...}{Optional arguments.}
}
\description{
Print method for \code{get_monte_carlo_se}-objects
}
