% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{print.plcp_multi}
\alias{print.plcp_multi}
\title{Print method for \code{study_parameters}-multiobjects}
\usage{
\method{print}{plcp_multi}(x, print_max = 10, empty = ".", digits = 2,
  ...)
}
\arguments{
\item{x}{An object of class \code{plcp_multi}.}

\item{print_max}{The number of rows to show}

\item{empty}{Symbol used to replace repeating non-unique parameters}

\item{digits}{Digits to show}

\item{...}{Optional arguments.}
}
\description{
Print method for \code{study_parameters}-multiobjects
}
