% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{print.plcp_multi_power}
\alias{print.plcp_multi_power}
\title{Print method for \code{get_power}-multi}
\usage{
\method{print}{plcp_multi_power}(x, ...)
}
\arguments{
\item{x}{An object of class \code{plcp_multi_power}.}

\item{...}{Optional arguments}
}
\description{
Print method for \code{get_power}-multi
}
