% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{print.plcp_multi_sim}
\alias{print.plcp_multi_sim}
\title{Print method for \code{simulate.plcp_multi}-objects}
\usage{
\method{print}{plcp_multi_sim}(x, ...)
}
\arguments{
\item{x}{An object created with \code{\link{simulate.plcp_multi}}}

\item{...}{Optional arguments.}
}
\description{
Print method for \code{simulate.plcp_multi}-objects
}
