% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{print.plcp_power_2lvl}
\alias{print.plcp_power_2lvl}
\title{Print method for two-level \code{get_power}}
\usage{
\method{print}{plcp_power_2lvl}(x, ...)
}
\arguments{
\item{x}{An object of class \code{plcp_power_2lvl}.}

\item{...}{Optional arguments}
}
\description{
Print method for two-level \code{get_power}
}
