% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPC.r
\name{print.plcp_sds}
\alias{print.plcp_sds}
\title{Print method for \code{get_sds}-objects}
\usage{
\method{print}{plcp_sds}(x, ...)
}
\arguments{
\item{x}{An object of class \code{plcp_sds}.}

\item{...}{Optional arguments.}
}
\description{
Print method for \code{get_sds}-objects
}
