% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{print.plcp_sim_formula}
\alias{print.plcp_sim_formula}
\alias{print.plcp_compare_sim_formula}
\title{Print method for simulation formulas}
\usage{
\method{print}{plcp_sim_formula}(x, ...)

\method{print}{plcp_compare_sim_formula}(x, ...)
}
\arguments{
\item{x}{A formula object.}

\item{...}{Not used}
}
\description{
Print method for simulation formulas
}
