% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{print.plcp_sim_summary}
\alias{print.plcp_sim_summary}
\title{Print method for \code{summary.plcp_sim}-objects}
\usage{
\method{print}{plcp_sim_summary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{plcp_sim_summary}}

\item{...}{Optional arguments.}
}
\description{
Print method for \code{summary.plcp_sim}-objects
}
