% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPC.r
\name{print.plcp_VPC}
\alias{print.plcp_VPC}
\title{Print method for \code{get_vpc}-objects}
\usage{
\method{print}{plcp_VPC}(x, digits = 2, ...)
}
\arguments{
\item{x}{Object created with \code{link{get_VPC}}}

\item{digits}{Number of digits to print}

\item{...}{Optional arguments}
}
\description{
Print method for \code{get_vpc}-objects
}
