% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{shiny_powerlmm}
\alias{shiny_powerlmm}
\title{Launch powerlmm's Shiny web application}
\usage{
shiny_powerlmm()
}
\description{
This Shiny application provides the basic functionality of the \pkg{powerlmm}-package
in a user friendly web application.
}
\examples{
\dontrun{

library(shiny)
shiny_powerlmm()
}
}
