% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdle_simulate_data.R
\name{simulate_data.plcp_hurdle}
\alias{simulate_data.plcp_hurdle}
\title{Generate Hurdle-lognormal data}
\usage{
\method{simulate_data}{plcp_hurdle}(paras, gen_fake = "sim_hurdle")
}
\arguments{
\item{paras}{paras}

\item{gen_fake}{func}
}
\value{
df
}
\description{
Generate Hurdle-lognormal data
}
