% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{[.plcp_multi_power}
\alias{[.plcp_multi_power}
\title{Subset function for \code{plcp_multi_power}-objects}
\usage{
\method{[}{plcp_multi_power}(x, i, ...)
}
\arguments{
\item{x}{A \code{plcp_multi_power}-object.}

\item{i}{Indicates which rows to subset.}

\item{...}{Ignored.}
}
\description{
Custom subset function for \code{plcp_multi_power}-object to make it compatible
with its print method.
}
