% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{transform_to_posttest}
\alias{transform_to_posttest}
\title{Helper to transform to posttest}
\usage{
transform_to_posttest(data)
}
\arguments{
\item{data}{}
}
\description{
Helper to transform to posttest
}
