% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_formula}
\alias{sim_formula}
\title{Create simulation formula}
\usage{
sim_formula(formula, data_transform = NULL, test = "time:treatment")
}
\arguments{
\item{formula}{A \code{character} containing a lme4 formula.}

\item{data_transform}{Optional; a \code{function} that applies a transformation
to the data during each simulation.}

\item{test}{A \code{character} vector indicating wich parameters should be tested.
Only applies to tests using Satterthwaite dfs, or when calculating confidence intervals.}
}
\description{
Create simulation formula
}
\examples{
f <- sim_formula("y ~ treatment * time + (1 + time | subject)")
}
